requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cffi==1.17.1
colorama==0.4.6
discord.py==2.4.0
exceptiongroup==1.2.2
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
soupsieve==2.6
tomli==2.2.1
typing_extensions==4.12.2
yarl==1.18.3
yt-dlp==2024.12.23
